refactor: extract session persistence from AgentWorkspace
